5.7.0

+++ src/pages/Mobile/Members/index.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react';
 import { connect } from 'react-redux';
 import * as actions from './redux/actions';
-import { WhiteSpace, Card, Flex, Modal, ActivityIndicator } from 'antd-mobile';
+import { Dialog, SpinLoading, Card, Input } from 'antd-mobile';
 import { Icon } from 'ming-ui';
 import Back from '../components/Back';
 import { ROLE_CONFIG, sysRoleType } from 'src/pages/Role/config.js';
@@ -15,7 +15,6 @@ class Members extends Component {
     this.state = {
       checked: this.props.memberData.rolesVisibleConfig === ROLE_CONFIG.REFUSE,
     };
-    this.modal = null;
   }
   componentDidMount() {
     $('html').addClass('mobileMembers');
@@ -24,51 +23,47 @@ class Members extends Component {
   }
   componentWillUnmount() {
     $('html').removeClass('mobileMembers');
-    if (this.modal) {
-      this.modal.close();
-    } else {
-      this.modal = null;
-    }
   }
   handleExitApp = () => {
     const { detail } = this.props.memberData;
-
-    this.modal = Modal.prompt(
-      <span className="Font17 bold">{_l('确认删除应用吗？')}</span>,
-      <span className="Font13 Gray mBottom5">
-        {_l('应用下所有配置和数据将被永久删除，不可恢复，请确认是否执行此操作！')}
-      </span>,
-      [
-        { text: _l('取消'), onPress: () => {}, style: { color: '#2196f3' } },
-        {
-          text: _l('确定'),
-          onPress: value => {
-            if (detail.name === value) {
-              this.props.dispatch(
-                actions.deleteApp(
-                  {
-                    projectId: '',
-                    appId: this.props.match.params.appId,
-                    isHomePage: true,
-                  },
-                  ({ data }) => {
-                    if (data) {
-                      this.props.history.push('/mobile/dashboard');
-                    }
-                  },
-                ),
-              );
-            } else {
-              alert(_l('应用名称错误'), 2);
-            }
-          },
-          style: { color: 'red' },
-        },
-      ],
-      'default',
-      null,
-      [_l('输入应用名称以确定删除')],
-    );
+    Dialog.confirm({
+      title: <span className="Font17 bold">{_l('确认删除应用吗？')}</span>,
+      content: (
+        <div className="TxtCenter">
+          <span className="Font13 Gray mBottom5">
+            {_l('应用下所有配置和数据将被永久删除，不可恢复，请确认是否执行此操作！')}
+          </span>
+          <Input
+            className="mTop10 pAll5 appNameInput"
+            style={{ borderRadius: 4, border: '1px solid #ededed', '--font-size': 13 }}
+            placeholder={_l('输入应用名称以确定删除')}
+          />
+        </div>
+      ),
+      cancelText: _l('取消'),
+      confirmText: <span className="Red">{_l('确认')}</span>,
+      onConfirm: () => {
+        const { value } = document.querySelector('.appNameInput input');
+        if (detail.name === value) {
+          this.props.dispatch(
+            actions.deleteApp(
+              {
+                projectId: '',
+                appId: this.props.match.params.appId,
+                isHomePage: true,
+              },
+              ({ data }) => {
+                if (data) {
+                  this.props.history.push('/mobile/dashboard');
+                }
+              },
+            ),
+          );
+        } else {
+          alert(_l('应用名称错误'), 2);
+        }
+      }
+    });
   };
   renderCard = (data, isAdmin) => {
     const { params } = this.props.match;
@@ -82,36 +77,32 @@ class Members extends Component {
 
       return (
         <Fragment key={item.roleId}>
-          <WhiteSpace size="md" />
           <Card
+            className="mTop10 Relative"
             onClick={() => {
               this.props.history.push(`/mobile/membersList/${params.appId}/${item.roleId}`);
             }}
           >
-            <Card.Body>
-              <div>
-                <Flex direction="row">
-                  <Flex.Item className="Gray Font17 Bold overflow_ellipsis">
-                    {isInCurrentRole && <span className="isMyRole" />}
-                    {item.label}
-                  </Flex.Item>
-                  <Flex.Item className="TxtRight moreAction">
-                    <span className="Gray_75 Bold TxtMiddle">{count > 0 && count}</span>
-                    <Icon icon="arrow-right-border" className="Font20 mLeft5 Gray_75 TxtMiddle" />
-                  </Flex.Item>
-                </Flex>
+            <div className="flexRow alignItemsCenter">
+              <div className="flex Gray Font17 Bold overflow_ellipsis">
+                {isInCurrentRole && <span className="isMyRole" />}
+                {item.label}
               </div>
-              <div className="Gray_75 Font14 mTop8 ellipsis">
-                {item.description ||
-                  (item.roleType === 100
-                    ? _l('可以配置应用，管理应用下所有数据和人员')
-                    : item.roleType === 2
-                    ? _l('管理所有数据和人员，不可配置应用')
-                    : item.roleType === 1
-                    ? _l('开发者只能配置应用')
-                    : _l('自定义权限'))}
+              <div className="flex TxtRight moreAction">
+                <span className="Gray_75 Bold TxtMiddle">{count > 0 && count}</span>
+                <Icon icon="arrow-right-border" className="Font20 mLeft5 Gray_75 TxtMiddle" />
               </div>
-            </Card.Body>
+            </div>
+            <div className="Gray_75 Font14 mTop8 ellipsis">
+              {item.description ||
+                (item.roleType === 100
+                  ? _l('可以配置应用，管理应用下所有数据和人员')
+                  : item.roleType === 2
+                  ? _l('管理所有数据和人员，不可配置应用')
+                  : item.roleType === 1
+                  ? _l('开发者只能配置应用')
+                  : _l('自定义权限'))}
+            </div>
           </Card>
         </Fragment>
       );
@@ -146,17 +137,14 @@ class Members extends Component {
             this.props.history.push(`/mobile/applyList/${params.appId}`);
           }}
         >
-          <Card.Body>
-            <Flex direction="row">
-              <Flex.Item className="Gray Font17 overflow_ellipsis pendingApply">{_l('待处理的申请')}</Flex.Item>
-              <Flex.Item className="TxtMiddle TxtRight moreAction">
-                <span className="Bold Red">{_l('%0人', memberData.applyList.length)}</span>
-                <Icon icon="arrow-right-border" className="Font20 mLeft5 Gray_75" />
-              </Flex.Item>
-            </Flex>
-          </Card.Body>
+          <div className="flexRow alignItemsCenter">
+            <div className="flex Gray Font17 overflow_ellipsis pendingApply">{_l('待处理的申请')}</div>
+            <div className="flex TxtMiddle TxtRight moreAction">
+              <span className="Bold Red">{_l('%0人', memberData.applyList.length)}</span>
+              <Icon icon="arrow-right-border" className="Font20 mLeft5 Gray_75" />
+            </div>
+          </div>
         </Card>
-        <WhiteSpace size="md" />
       </div>
     );
   }
@@ -170,11 +158,10 @@ class Members extends Component {
 
     return (
       <Fragment>
-        <WhiteSpace size="lg" />
-        <Card>
-          <Card.Body className="TxtCenter Red Font15 Bold" onClick={this.handleExitApp}>
+        <Card className="mTop10">
+          <div className="TxtCenter Red Font15 Bold" onClick={this.handleExitApp}>
             {_l('删除应用并退出')}
-          </Card.Body>
+          </div>
         </Card>
       </Fragment>
     );
@@ -184,9 +171,9 @@ class Members extends Component {
     const { params } = this.props.match;
     if (isMemberLoading) {
       return (
-        <Flex className="h100" justify="center" align="center">
-          <ActivityIndicator size="large" />
-        </Flex>
+        <div className="flexRow justifyContentCenter alignItemsCenter h100">
+          <SpinLoading color='primary' />
+        </div>
       );
     } else {
       const { detail, rolesVisibleConfig } = this.props.memberData;
